Create snakeBFS
This is a traditional Snake game with an interface made with GTK (reused library from paint.c). The twist is that if you hold the "i" key, the game will "solve" itself using BFS. This isn't optimal, and sometimes this causes the snake to lose in some silly ways, but I wanted to use this to practice data structures.